refactor: using postgres now

+++ internal/lib/paginated-list.go
@@ -2,6 +2,6 @@ package lib
 
 type PaginatedList[T any] struct {
 	Items      []T `json:"items"`
-	Page       int64 `json:"page"`
-	TotalPages int64 `json:"total_pages"`
+	Page       int32 `json:"page"`
+	TotalPages int32 `json:"total_pages"`
 }
